left operand refactor
